Autorizacion y Autenticacion finalizada

+++ app/Http/Controllers/Api/RecipeController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreRecipeRequest;
 use App\Http\Requests\UpdateRecipeRequest;
 
+use Illuminate\Support\Facades\Gate;
 
 use App\Models\Recipe;
 
@@ -22,7 +23,12 @@ class RecipeController extends Controller
     }
 
     public function store(StoreRecipeRequest $request){
-        $recipe = Recipe::create($request -> all());
+        $recipe = $request->user()->recipes()->create($request -> all());
+        $recipe ->tags()->attach(json_decode($request->tags));
+
+        $recipe -> image = $request-> file('image')->store('recipe', 'public');
+        $recipe ->save();
+        
         return response()->json(new RecipesResource($recipe), Response::HTTP_CREATED );
     }
 
@@ -32,14 +38,20 @@ class RecipeController extends Controller
     }
 
     public function update(UpdateRecipeRequest $request, Recipe $recipe){
+        Gate::authorize('update', $recipe);
         $recipe -> update(($request -> all()));
         if($tags = json_decode($request->tags)){
             $recipe->tags()->sync($tags);
+        }
+        if($request->file('image')){
+            $recipe -> image = $request-> file('image')->store('recipe', 'public');
+            $recipe ->save();
         }
         return response()->json(new RecipesResource($recipe), Response::HTTP_OK);
     }
 
     public function destroy(Recipe $recipe){
+        Gate::authorize('delete', $recipe);
         $recipe->delete();
         return response()->json(null, Response::HTTP_NO_CONTENT);
     }
